refactor: Extract Dimens & Typography objects to Compose:Basics module
