+++ Compose/Basics/src/main/kotlin/com/infomaniak/core/compose/basics/Typography.kt
@@ -15,14 +15,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.core.crossapplogin.front.theme
+package com.infomaniak.core.compose.basics
 
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 
-internal object Typography {
+object Typography {
 
     val h1 = TextStyle(
         fontFamily = FontFamily.Default,
