+++ CrossAppLogin/Front/src/main/kotlin/com/infomaniak/core/crossapplogin/front/components/AddAccount.kt
@@ -39,13 +39,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import com.infomaniak.core.compose.basics.Dimens
+import com.infomaniak.core.compose.basics.Typography
 import com.infomaniak.core.compose.margin.Margin
 import com.infomaniak.core.crossapplogin.front.R
 import com.infomaniak.core.crossapplogin.front.data.CrossLoginCustomization
 import com.infomaniak.core.crossapplogin.front.data.CrossLoginDefaults
 import com.infomaniak.core.crossapplogin.front.icons.AddUser
-import com.infomaniak.core.crossapplogin.front.theme.Dimens
-import com.infomaniak.core.crossapplogin.front.theme.Typography
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
