+++ CrossAppLogin/Front/src/main/kotlin/com/infomaniak/core/crossapplogin/front/components/BottomSheetItem.kt
@@ -41,13 +41,13 @@ import androidx.compose.ui.semantics.toggleableState
 import androidx.compose.ui.state.ToggleableState
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import com.infomaniak.core.compose.basics.Dimens
 import com.infomaniak.core.compose.margin.Margin
+import com.infomaniak.core.crossapplogin.back.ExternalAccount
 import com.infomaniak.core.crossapplogin.front.data.CrossLoginCustomization
 import com.infomaniak.core.crossapplogin.front.data.CrossLoginDefaults
 import com.infomaniak.core.crossapplogin.front.icons.Checkmark
 import com.infomaniak.core.crossapplogin.front.previews.AccountsPreviewParameter
-import com.infomaniak.core.crossapplogin.front.theme.Dimens
-import com.infomaniak.core.crossapplogin.back.ExternalAccount
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
