+++ CrossAppLogin/Front/src/main/kotlin/com/infomaniak/core/crossapplogin/front/components/ThreeAccountsView.kt
@@ -30,10 +30,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import com.infomaniak.core.compose.basics.Dimens
+import com.infomaniak.core.crossapplogin.back.ExternalAccount
 import com.infomaniak.core.crossapplogin.front.data.CrossLoginDefaults
 import com.infomaniak.core.crossapplogin.front.previews.AccountsPreviewParameter
-import com.infomaniak.core.crossapplogin.front.theme.Dimens
-import com.infomaniak.core.crossapplogin.back.ExternalAccount
+import com.infomaniak.core.crossapplogin.front.theme.Dimens as CrossLoginDimens
 
 @Composable
 internal fun ThreeAccountsView(
@@ -50,7 +51,7 @@ internal fun ThreeAccountsView(
                 strokeColor = avatarStrokeColor,
             )
 
-            Spacer(Modifier.width(width = Dimens.avatarsBoxWidth - (Dimens.iconSize + Dimens.iconSize)))
+            Spacer(Modifier.width(width = CrossLoginDimens.avatarsBoxWidth - (Dimens.iconSize + Dimens.iconSize)))
 
             // Right
             CrossLoginAvatar(
@@ -62,7 +63,7 @@ internal fun ThreeAccountsView(
 
         // Center
         CrossLoginAvatar(
-            modifier = Modifier.size(Dimens.avatarsBoxHeight),
+            modifier = Modifier.size(CrossLoginDimens.avatarsBoxHeight),
             account = accounts[0],
             strokeColor = avatarStrokeColor,
         )
