+++ kSuite/MykSuite/src/main/java/com/infomaniak/core/ksuite/myksuite/ui/components/TextWithIcon.kt
@@ -40,7 +40,7 @@ import com.infomaniak.core.compose.margin.Margin
 import com.infomaniak.core.ksuite.myksuite.R
 import com.infomaniak.core.ksuite.myksuite.ui.theme.LocalMyKSuiteColors
 import com.infomaniak.core.ksuite.myksuite.ui.theme.MyKSuiteTheme
-import com.infomaniak.core.ksuite.myksuite.ui.theme.Typography
+import com.infomaniak.core.compose.basics.Typography
 
 /** This component allows to put an icon on any line of the text, thanks to the [iconLine] parameter
  *  This code comes from [here](https://stackoverflow.com/questions/70708056/how-to-centrally-align-icon-to-first-line-of-a-text-component-in-compose/71312465#71312465)
