+++ kSuite/MykSuite/src/main/java/com/infomaniak/core/ksuite/myksuite/ui/theme/Theme.kt
@@ -29,6 +29,7 @@ import androidx.compose.runtime.ProvidableCompositionLocal
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.graphics.Color
 import com.google.android.material.color.MaterialColors
+import com.infomaniak.core.compose.basics.Typography
 import com.infomaniak.core.compose.materialthemefromxml.MaterialThemeFromXml
 
 internal val LocalMyKSuiteColors: ProvidableCompositionLocal<MyKSuiteColors> = staticCompositionLocalOf { MyKSuiteColors() }
